review some tests #3620
