+++ tests/Functional/Controller/Api/VisiteCreateControllerTest.php
@@ -2,7 +2,6 @@
 
 namespace App\Tests\Functional\Controller\Api;
 
-use App\Entity\Intervention;
 use App\Entity\User;
 use App\Repository\SignalementRepository;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
@@ -12,7 +11,7 @@ use Symfony\Component\Routing\RouterInterface;
 
 class VisiteCreateControllerTest extends WebTestCase
 {
-    public const string UUID_SIGNALEMENT = '00000000-0000-0000-2022-000000000006';
+    public const string UUID_SIGNALEMENT = '00000000-0000-0000-2023-000000000026';
 
     private KernelBrowser $client;
     private RouterInterface $router;
@@ -29,28 +28,15 @@ class VisiteCreateControllerTest extends WebTestCase
         $this->client->loginUser($user, 'api');
     }
 
-    /** @dataProvider provideData */
-    public function testCreateVisite(string $signalementUuid, bool $notifyUsager, int $nbMailSent): void
+    /** @dataProvider provideDataForNotification */
+    public function testCreateVisiteWithNotification(string $type, array $payload, int $nbMailSent): void
     {
-        $signalement = self::getContainer()->get(SignalementRepository::class)->findOneBy(['uuid' => $signalementUuid]);
+        $signalement = self::getContainer()->get(SignalementRepository::class)->findOneBy(['uuid' => self::UUID_SIGNALEMENT]);
         $firstFile = $signalement->getFiles()->first();
         $lastFile = $signalement->getFiles()->last();
-        $payload = [
-            'date' => '2025-01-01',
-            'time' => '12:00',
-            'occupantPresent' => true,
-            'proprietairePresent' => true,
-            'notifyUsager' => $notifyUsager,
-            'concludeProcedure' => [
-                'LOGEMENT_DECENT',
-                'RESPONSABILITE_OCCUPANT_ASSURANTIEL',
-            ],
-            'details' => 'lorem ipsum dolor sit <em>amet</em>',
-            'files' => [
-                $firstFile->getUuid(), $lastFile->getUuid(),
-            ],
-        ];
-
+        if ('visite_confirmed' === $type) {
+            $payload['files'] = [$firstFile->getUuid(), $lastFile->getUuid()];
+        }
         $this->client->request(
             method: 'POST',
             uri: $this->router->generate('api_signalements_visite_post', ['uuid' => self::UUID_SIGNALEMENT]),
@@ -61,17 +47,161 @@ class VisiteCreateControllerTest extends WebTestCase
         $this->assertEquals(201, $this->client->getResponse()->getStatusCode());
         $this->assertEmailCount($nbMailSent);
 
-        /** @var Intervention $lastIntervention */
-        $lastIntervention = $signalement->getInterventions()->last();
-
-        $crawler = new Crawler($lastIntervention->getDetails());
-        $links = $crawler->filter('a.fr-link');
-        $this->assertCount(2, $links, 'Il doit y avoir exactement 2 liens dans le contenu HTML.');
+        if ('visite_confirmed' === $type) {
+            $content = json_decode($this->client->getResponse()->getContent(), true);
+            $crawler = new Crawler($content['details']);
+            $links = $crawler->filter('a.fr-link');
+            $this->assertCount(2, $links, 'Il doit y avoir exactement 2 liens dans le contenu HTML.');
+        }
     }
 
-    public function provideData(): \Generator
+    /** @dataProvider provideDataForPendingVisite */
+    public function testCreateVisiteWithPendingVisiteWithErrors(array $payload): void
     {
-        yield 'test create visite with usager notification' => [self::UUID_SIGNALEMENT, true, 2];
-        yield 'test create visite with no usager notification' => [self::UUID_SIGNALEMENT, false, 1];
+        $signalementUuid = '00000000-0000-0000-2022-000000000006';
+        $signalement = self::getContainer()->get(SignalementRepository::class)->findOneBy(['uuid' => $signalementUuid]);
+        $this->client->request(
+            method: 'POST',
+            uri: $this->router->generate('api_signalements_visite_post', ['uuid' => $signalementUuid]),
+            server: ['CONTENT_TYPE' => 'application/json'],
+            content: json_encode($payload)
+        );
+
+        $this->assertEquals(400, $this->client->getResponse()->getStatusCode());
+        $message = json_decode($this->client->getResponse()->getContent(), true)['message'];
+        $this->assertStringContainsString($signalement->getInterventions()->first()->getUuid(), $message);
+    }
+
+    /** @dataProvider provideDataForPendingVisite */
+    public function testCreateVisiteWithPendingVisite(array $payload): void
+    {
+        $signalementUuid = '00000000-0000-0000-2022-000000000006';
+        $signalement = self::getContainer()->get(SignalementRepository::class)->findOneBy(['uuid' => $signalementUuid]);
+        $this->client->request(
+            method: 'POST',
+            uri: $this->router->generate('api_signalements_visite_post', ['uuid' => $signalementUuid]),
+            server: ['CONTENT_TYPE' => 'application/json'],
+            content: json_encode($payload)
+        );
+
+        $this->assertEquals(400, $this->client->getResponse()->getStatusCode());
+        $message = json_decode($this->client->getResponse()->getContent(), true)['message'];
+        $this->assertStringContainsString($signalement->getInterventions()->first()->getUuid(), $message);
+    }
+
+    /** @dataProvider  provideDataErrorPayload */
+    public function testCreateVisiteWithPayloadErrors(array $payload, array $fieldsErrors, string $errorMessage): void
+    {
+        $this->client->request(
+            method: 'POST',
+            uri: $this->router->generate('api_signalements_visite_post', ['uuid' => self::UUID_SIGNALEMENT]),
+            server: ['CONTENT_TYPE' => 'application/json'],
+            content: json_encode($payload)
+        );
+
+        $errors = json_decode($this->client->getResponse()->getContent(), true)['errors'];
+        $this->assertStringContainsString($errorMessage, $errors[0]['message']);
+        $errors = array_map(function ($error) { return $error['property']; }, $errors);
+        $this->assertEquals(400, $this->client->getResponse()->getStatusCode());
+        $this->assertEquals($fieldsErrors, $errors);
+    }
+
+    public function provideDataForNotification(): \Generator
+    {
+        yield 'test create visite confirmed with usager notification' => [
+            'visite_confirmed',
+            [
+                'date' => '2025-01-01',
+                'time' => '12:00',
+                'occupantPresent' => true,
+                'proprietairePresent' => true,
+                'notifyUsager' => true,
+                'concludeProcedure' => [
+                    'LOGEMENT_DECENT',
+                    'RESPONSABILITE_OCCUPANT_ASSURANTIEL',
+                ],
+                'details' => 'lorem ipsum dolor sit <em>amet</em>',
+            ],
+            2,
+        ];
+        yield 'test create visite confirmed with no usager notification' => [
+            'visite_confirmed',
+            [
+                'date' => '2025-01-01',
+                'time' => '12:00',
+                'occupantPresent' => true,
+                'proprietairePresent' => true,
+                'notifyUsager' => false,
+                'concludeProcedure' => [
+                    'LOGEMENT_DECENT',
+                    'RESPONSABILITE_OCCUPANT_ASSURANTIEL',
+                ],
+                'details' => 'lorem ipsum dolor sit <em>amet</em>',
+            ],
+            1,
+        ];
+
+        yield 'test create visite planned' => [
+            'visite_planned',
+            [
+                'date' => '2125-01-01',
+                'time' => '12:00',
+            ],
+            1,
+        ];
+    }
+
+    public function provideDataForPendingVisite(): \Generator
+    {
+        yield 'test create visite planned' => [
+            [
+                'date' => '2200-06-01',
+                'time' => '12:00',
+            ],
+        ];
+        yield 'test create visite confirmed' => [
+            [
+                'date' => '2024-06-01',
+                'time' => '12:00',
+                'occupantPresent' => true,
+                'proprietairePresent' => true,
+                'notifyUsager' => false,
+                'details' => 'lorem ipsum dolor sit <em>amet</em>',
+                'concludeProcedure' => [
+                    'LOGEMENT_DECENT',
+                ],
+            ],
+        ];
+    }
+
+    public function provideDataErrorPayload(): \Generator
+    {
+        yield 'test create visite confirmed with missing data' => [
+            [
+                'date' => '2020-06-01',
+                'time' => '12:00',
+            ],
+            [
+                'occupantPresent', 'proprietairePresent', 'notifyUsager', 'concludeProcedure', 'details',
+            ],
+            'est obligatoire pour une visite effectuée',
+        ];
+        yield 'test create visite planned with more data than expected' => [
+            [
+                'date' => '2126-06-01',
+                'time' => '12:00',
+                'occupantPresent' => true,
+                'proprietairePresent' => true,
+                'notifyUsager' => false,
+                'details' => 'lorem ipsum dolor sit <em>amet</em>',
+                'concludeProcedure' => [
+                    'LOGEMENT_DECENT',
+                ],
+            ],
+            [
+                'occupantPresent', 'proprietairePresent', 'notifyUsager', 'concludeProcedure', 'details',
+            ],
+            'ne peut être renseigné que si la visite a été effectuée',
+        ];
     }
 }
